Splitting up components of the simulation that can be treated independently of the each other.  There is a game portion in game.py, and the simulation itself now in simulation.py.
