+++ run.py
@@ -9,6 +9,7 @@ import os
 import numpy as np
 from camera import Camera
 from utils import clean_filename
+import simulation
 
 logs_directory = '/tmp/gravipy_log' or os.path.join(os.path.dirname(os.path.realpath(__file__)), 'log')
 
@@ -37,31 +38,31 @@ game.log.addHandler(fh)
 game.body.log.addHandler(ch)
 game.body.log.addHandler(fh)
 
+
 Camera.log.addHandler(ch)
 Camera.log.addHandler(fh)
 
 config = {
     "dimensions": (1680, 900),
-    "gravitational_constant": 0.1,
-    "draw_sphere_of_influence": False,
+    "gravitational_constant": 0.5,
+    "draw_sphere_of_influence": True,
     "num_bg_stars": 0,
     "enable_movement": False}
 
 black = 0, 0, 0
 
-
 pygame.init()
 
-planets1 = game.generate_star_system_config("Sol", (1000, 1000, -4000), 1)
-planets2 = game.generate_star_system_config("Sol", (-15000, 1000, 0), 2)
-planets3 = game.generate_star_system_config("Sol", (8000, -10000, 4000), 1)
-planets4 = game.generate_star_system_config("Sol", (-8000, 1000, 8000), 1)
-planets5 = game.generate_star_system_config("Sol", (4000, -4000, 4000), 1)
+planets1 = simulation.generate_star_system_config("Sol", (1000, 1000, -4000), 1)
+planets2 = simulation.generate_star_system_config("Sol", (-15000, 1000, 0), 2)
+planets3 = simulation.generate_star_system_config("Sol", (8000, -10000, 4000), 1)
+planets4 = simulation.generate_star_system_config("Sol", (-8000, 1000, 8000), 1)
+planets5 = simulation.generate_star_system_config("Sol", (4000, -4000, 4000), 1)
 planets = planets1 + planets2 + planets3 + planets4 + planets5
 
-planets2 = game.generate_star_system_config("Sol", (10, 10, 0), 5)
+planets2 = simulation.generate_star_system_config("Sol", (10, 10, 0), 5)
 
-sim = game.GravitySimulation(planets2, config)
+sim = game.GravitySimulationSystem(planets2, config)
 cam = Camera(np.array([0, -5000, 300]), config["dimensions"])
 screen = pygame.display.set_mode(config["dimensions"])
 background = pygame.Surface(config["dimensions"])
@@ -69,44 +70,10 @@ background = pygame.Surface(config["dimensions"])
 background.fill(black)
 sim.draw_background(background, cam)
 
-pause = True
-textcolor = 128, 128, 64
-font = pygame.font.Font(None, 48)
-pause_image = font.render("PRESS 'P' '<' or '>' TO UNPAUSE", True, textcolor, black)
-
-
-timestep = 1
-timewarp_value = 1
-max_timewarp = 8
-arrow_size = 64, 64
-timewarp_arrow_color = 96, 64, 96
-timewarp_image_bgcolor = 255, 196, 255
-arrow_image = pygame.Surface(arrow_size)
-arrow_image.fill(timewarp_image_bgcolor)
-pygame.draw.polygon(arrow_image, timewarp_arrow_color, [(8, 8), (56, 32), (8, 56)])
-
-def get_timewarp_image():
-    image_size = timewarp_value * arrow_size[0], arrow_size[1]
-    image = pygame.Surface(image_size)
-
-    for n in range(timewarp_value):
-        image.blit(arrow_image, (n*arrow_size[0], 0))
-
-    return image
-
-
-timewarp_image = get_timewarp_image()
-
-
 pygame.mixer.init()
 music = pygame.mixer.Sound(file="./sound/spheric_lounge_books_of_mantra.ogg")
 music.play()
 
-
-pause = True
-textcolor = 128, 128, 64
-font = pygame.font.Font(None, 48)
-pause_image = font.render("PRESS 'P' '<' or '>' TO UNPAUSE", True, textcolor, black)
 
 clock = pygame.time.Clock()
 clock.tick()
@@ -119,32 +86,6 @@ while 1:
         if event.type == pygame.QUIT:
             sys.exit()
 
-        if event.type == pygame.KEYDOWN:
-            if event.key == pygame.K_SPACE:
-                log.info("Restarting simulation")
-                sim.reset()
-                cam.reset()
-
-            if event.key == pygame.K_p:
-                pause = not pause
-
-            if event.key == pygame.K_COMMA:
-                if 0 < timewarp_value:
-                    timewarp_value -= 1
-                    pause = False
-                    timewarp_image = get_timewarp_image()
-                if timewarp_value == 0:
-                    pause = True
-
-            if event.key == pygame.K_PERIOD:
-                if timewarp_value < max_timewarp:
-                    timewarp_value += 1
-                    pause = False
-                    timewarp_image = get_timewarp_image()
-
-            if event.key == pygame.K_F12:
-                pygame.display.toggle_fullscreen()
-
         cam.handle_event(event)
         if event.type == Camera.CAMERAEVENT and event.movement == Camera.CAMERATURNED:
             background.fill(black)
@@ -155,14 +96,8 @@ while 1:
     screen.fill(black)
     screen.blit(background, (0, 0))
 
-    if pause is False:
-        sim.update_planets(2 ** (timewarp_value - 1))
-        screen.blit(timewarp_image, (100, 800))
-    else:
-        screen.blit(pause_image, (100, 100))
-
-    sim.update_sim(cam)
-    sim.draw_planets(screen, cam)
+    sim.step()
+    sim.draw(screen, cam)
     pygame.display.flip()
 
     clock.tick(30)
